remove adding bias from fp8 kernel, and small refactors
