+++ src/nanotron/fp8/dtypes.py
@@ -1,0 +1,7 @@
+from enum import Enum, auto
+
+
+class DTypes(Enum):
+    FP8E4M3 = auto()
+    FP8E5M2 = auto()
+    kfloat16 = auto()
