+++ tests/fp8/test_linear.py
@@ -1,7 +1,7 @@
 import pytest
 import torch
 import transformer_engine as te  # noqa
-from nanotron.fp8.constants import DTypes
+from nanotron.fp8.dtypes import DTypes
 from nanotron.fp8.linear import FP8Linear
 from nanotron.fp8.parameter import FP8Parameter
 from nanotron.fp8.tensor import FP8Tensor
