+++ tests/fp8/test_tensor.py
@@ -5,7 +5,7 @@ import pytest
 import torch
 import transformer_engine as te  # noqa
 import transformer_engine_extensions as tex
-from nanotron.fp8.constants import DTypes
+from nanotron.fp8.dtypes import DTypes
 from nanotron.fp8.meta import FP8Meta
 from nanotron.fp8.tensor import FP8Tensor, convert_tensor_from_fp8
